+++ language-server/src/engine.rs
@@ -31,6 +31,8 @@ use lsp_types::{
 };
 use std::{collections::HashSet, sync::Arc};
 
+use crate::rename::rename_module_alias;
+
 use super::{
     DownloadDependencies, MakeLocker,
     code_action::{
@@ -58,6 +60,7 @@ use super::{
     rename::{RenameOutcome, RenameTarget, Renamed, rename_local_variable, rename_module_entity},
     signature_help, src_span_to_lsp_range,
 };
+
 #[derive(Debug, PartialEq, Eq)]
 pub struct Response<T> {
     pub result: Result<T, Error>,
@@ -643,7 +646,9 @@ where
 
             let byte_index = lines.byte_index(params.position);
 
-            Ok(match reference_for_ast_node(found, &current_module.name) {
+            let referenced = reference_for_ast_node(found, &current_module.name);
+
+            Ok(match referenced {
                 Some(Referenced::LocalVariable {
                     location, origin, ..
                 }) if location.contains(byte_index) => match origin.map(|origin| origin.syntax) {
@@ -685,6 +690,8 @@ where
                         None
                     }
                 }
+                Some(Referenced::ModuleName { location, .. }) => success_response(location),
+
                 _ => None,
             })
         })
@@ -706,7 +713,9 @@ where
                 return Ok(RenameOutcome::NoRenames.into_result());
             };
 
-            Ok(match reference_for_ast_node(found, &module.name) {
+            let referenced = reference_for_ast_node(found, &module.name);
+
+            Ok(match referenced {
                 Some(Referenced::LocalVariable {
                     origin,
                     definition_location,
@@ -776,6 +785,13 @@ where
                 )
                 .into_result(),
 
+                Some(Referenced::ModuleName {
+                    module_name,
+                    module_alias,
+                    ..
+                }) => rename_module_alias(module, &lines, &params, &module_name, &module_alias)
+                    .into_result(),
+
                 None => RenameOutcome::NoRenames.into_result(),
             })
         })
@@ -801,7 +817,9 @@ where
 
             let byte_index = lines.byte_index(position.position);
 
-            Ok(match reference_for_ast_node(found, &module.name) {
+            let referenced = reference_for_ast_node(found, &module.name);
+
+            Ok(match referenced {
                 Some(Referenced::LocalVariable {
                     origin,
                     definition_location,
@@ -1018,8 +1036,12 @@ Unused labelled fields:
                 Located::Label(location, type_) => {
                     Some(hover_for_label(location, type_, lines, module))
                 }
-                Located::ModuleName { location, name, .. } => {
-                    let Some(module) = this.compiler.get_module_interface(name) else {
+                Located::ModuleName {
+                    location,
+                    module_name,
+                    ..
+                } => {
+                    let Some(module) = this.compiler.get_module_interface(&module_name) else {
                         return Ok(None);
                     };
                     Some(hover_for_module(module, location, &lines, &this.hex_deps))
