+++ language-server/src/tests/rename.rs
@@ -23,8 +23,10 @@ fn rename(
         engine.prepare_rename(params).result.unwrap()
     });
 
-    let Some(lsp_types::PrepareRenameResponse::Range(range)) = prepare_rename_response else {
-        return Ok(None);
+    let range = match prepare_rename_response {
+        Some(lsp_types::PrepareRenameResponse::Range(range)) => range,
+        Some(lsp_types::PrepareRenameResponse::RangeWithPlaceholder { range, .. }) => range,
+        _ => return Ok(None),
     };
 
     let outcome = tester.at(position, |engine, params, _| {
@@ -1752,6 +1754,16 @@ pub fn go(x) {
 }
 
 #[test]
+fn rename_module_from_import() {
+    assert_rename!(
+        TestProject::for_source("import      option")
+            .add_module("option", "pub type Option(a) { Some(a) None }"),
+        "opt",
+        find_position_of("option"),
+    );
+}
+
+#[test]
 fn rename_works_when_error_is_present() {
     assert_rename!(
         r#"
@@ -1766,6 +1778,102 @@ pub fn main() {
   "#,
         "wobble",
         find_position_of("wibble")
+    );
+}
+
+#[test]
+fn rename_module_from_import_with_alias() {
+    assert_rename!(
+        TestProject::for_source("import   option    as      opt")
+            .add_module("option", "pub type Option(a) { Some(a) None }"),
+        "o",
+        find_position_of("opt"),
+    );
+}
+
+#[test]
+fn reanem_module_from_import_with_unqualified_values() {
+    assert_rename!(
+        TestProject::for_source("import   option   .   {    Some, None }")
+            .add_module("option", "pub type Option(a) { Some(a) None }"),
+        "opt",
+        find_position_of("option"),
+    );
+}
+
+#[test]
+fn rename_module_from_import_with_unqualified_value_and_alias() {
+    assert_rename!(
+        TestProject::for_source("import   option  .     {Some, None}   as     opt")
+            .add_module("option", "pub type Option(a) { Some(a) None }"),
+        "o",
+        find_position_of("opt"),
+    );
+}
+
+#[test]
+fn rename_module_from_import_namespaced() {
+    assert_rename!(
+        TestProject::for_source("import   std /   option  ")
+            .add_module("std/option", "pub type Option(a) { Some(a) None }"),
+        "opt",
+        find_position_of("option"),
+    );
+}
+
+#[test]
+fn rename_module_from_import_namespaced_with_alias() {
+    assert_rename!(
+        TestProject::for_source("import   std /   option   as     opt")
+            .add_module("std/option", "pub type Option(a) { Some(a) None }"),
+        "o",
+        find_position_of("opt"),
+    );
+}
+
+#[test]
+fn rename_module_from_import_namespaced_with_unqualified_values() {
+    assert_rename!(
+        TestProject::for_source("import   std /   option   .   {    Some, None }")
+            .add_module("std/option", "pub type Option(a) { Some(a) None }"),
+        "opt",
+        find_position_of("option"),
+    );
+}
+
+#[test]
+fn rename_module_from_import_namespaced_with_unqualified_value_and_alias() {
+    assert_rename!(
+        TestProject::for_source("import   std /   option   .   {    Some, None }   as     opt")
+            .add_module("std/option", "pub type Option(a) { Some(a) None }"),
+        "o",
+        find_position_of("opt"),
+    );
+}
+
+#[test]
+fn rename_module_from_import_with_alias_to_original_name() {
+    assert_rename!(
+        TestProject::for_source("import   option   as     opt")
+            .add_module("option", "pub type Option(a) { Some(a) None }"),
+        "option",
+        find_position_of("opt"),
+    );
+}
+
+#[test]
+fn rename_module_from_variant_in_expression() {
+    let src = r#"
+import option
+
+pub fn main() {
+  echo option     . None
+}
+"#;
+    assert_rename!(
+        TestProject::for_source(src).add_module("option", "pub type Option(a) { Some(a) None }"),
+        "opt",
+        find_position_of("option").nth_occurrence(2)
     );
 }
 
@@ -1787,6 +1895,73 @@ fn main() -> String {
 }
 
 #[test]
+fn rename_module_from_constant_in_expression() {
+    let src = r#"
+import maths
+
+pub fn main() {
+  echo maths  .    pi
+}
+"#;
+
+    assert_rename!(
+        TestProject::for_source(src).add_module("maths", "pub const pi = 3.14"),
+        "m",
+        find_position_of("maths").nth_occurrence(2),
+    );
+}
+
+#[test]
+fn rename_module_from_variant_in_const() {
+    let src = r#"
+import option
+
+const x = option   .None
+"#;
+
+    assert_rename!(
+        TestProject::for_source(src).add_module("option", "pub type Option(a) { Some(a) None }"),
+        "opt",
+        find_position_of("option").nth_occurrence(2),
+    );
+}
+
+#[test]
+fn rename_module_from_constant_in_const() {
+    let src = r#"
+import maths
+
+const x = maths   .  pi
+"#;
+
+    assert_rename!(
+        TestProject::for_source(src).add_module("maths", "pub const pi = 3.14"),
+        "m",
+        find_position_of("maths").nth_occurrence(2),
+    );
+}
+
+#[test]
+fn rename_module_from_variant_in_pattern() {
+    let src = r#"
+import option
+
+pub fn is_some(option) {
+  case option {
+    option.  Some(_) -> True
+    option  .None -> False
+  }
+}
+"#;
+
+    assert_rename!(
+        TestProject::for_source(src).add_module("option", "pub type Option(a) { Some(a) None }"),
+        "opt",
+        find_position_of("option").nth_occurrence(4),
+    );
+}
+
+#[test]
 fn rename_prefix_string_suffix_variable_in_case_triggered_from_usage() {
     assert_rename!(
         "
@@ -1800,6 +1975,27 @@ fn main() -> String {
 ",
         "new_name",
         find_position_of("rest").nth_occurrence(2)
+    );
+}
+
+#[test]
+fn rename_module_from_variant_in_clause_guard() {
+    let src = r#"
+import option
+
+pub fn count_none(list) {
+  case list {
+    [option, ..rest] if option == option  . None -> 1 + count_none(rest)
+    [_, ..rest] -> count_none(rest)
+    [] -> 0
+  }
+}
+"#;
+
+    assert_rename!(
+        TestProject::for_source(src).add_module("option", "pub type Option(a) { Some(a) None }"),
+        "opt",
+        find_position_of("option").nth_occurrence(4),
     );
 }
 
@@ -1821,6 +2017,27 @@ fn main() -> String {
 }
 
 #[test]
+fn rename_module_from_constant_in_clause_guard() {
+    let src = r#"
+import maths
+
+pub fn count_pi(list) {
+  case list {
+    [number, ..rest] if number == maths  . pi -> 1 + count_pi(rest)
+    [_, ..rest] -> count_pi(rest)
+    [] -> 0
+  }
+}
+"#;
+
+    assert_rename!(
+        TestProject::for_source(src).add_module("maths", "pub const pi = 3.14"),
+        "m",
+        find_position_of("maths").nth_occurrence(2),
+    );
+}
+
+#[test]
 fn rename_prefix_string_suffix_variable_with_alternative_definition_triggered_from_second_pattern()
 {
     assert_rename!(
@@ -1835,6 +2052,85 @@ fn main() -> String {
 ",
         "new_name",
         find_position_of("rest").nth_occurrence(2),
+    );
+}
+
+#[test]
+fn rename_module_from_type_in_custom_type() {
+    let src = r#"
+import option
+
+type Value(a) {
+  Value(option.Option(a))
+}
+"#;
+
+    assert_rename!(
+        TestProject::for_source(src).add_module("option", "pub type Option(a) { Some(a) None }"),
+        "opt",
+        find_position_of("option").nth_occurrence(2),
+    );
+}
+
+#[test]
+fn rename_module_from_type_in_type_alias() {
+    let src = r#"
+import option
+
+type Option(a) =
+  option.Option(a)
+"#;
+
+    assert_rename!(
+        TestProject::for_source(src).add_module("option", "pub type Option(a) { Some(a) None }"),
+        "opt",
+        find_position_of("option").nth_occurrence(2),
+    );
+}
+
+#[test]
+fn rename_module_from_type_in_annotation() {
+    let src = r#"
+import option
+
+const x: option.Option(Int) = option.Some(1)
+"#;
+
+    assert_rename!(
+        TestProject::for_source(src).add_module("option", "pub type Option(a) { Some(a) None }"),
+        "opt",
+        find_position_of("option").nth_occurrence(2),
+    );
+}
+
+#[test]
+fn rename_module_from_function_call() {
+    let src = r#"
+import option
+
+pub fn main() {
+  option.is_some(option.Some(1))
+}
+"#;
+
+    let option_module = r#"
+pub type Option(a) {
+  Some(a)
+  None
+}
+
+pub fn is_some(option: Option(a)) -> Bool {
+  case option {
+    Some(_) -> True
+    None -> False
+  }
+}
+"#;
+
+    assert_rename!(
+        TestProject::for_source(src).add_module("option", option_module),
+        "opt",
+        find_position_of("option").nth_occurrence(2),
     );
 }
 
@@ -2040,5 +2336,22 @@ fn main() {
 ",
         "new_name",
         find_position_of("digit").nth_occurrence(1)
+    );
+}
+
+#[test]
+fn rename_module_from_alias_use() {
+    let src = r#"
+import maths  as     m
+
+pub fn main() {
+  echo m      .pi
+}
+"#;
+
+    assert_rename!(
+        TestProject::for_source(src).add_module("maths", "pub const pi = 3.14"),
+        "mth",
+        find_position_of("m").nth_occurrence(5)
     );
 }
